fix: Decomposed elements in  productTypes to different files.
