+++ src/types/product/buyByTypes.ts
@@ -1,0 +1,6 @@
+export enum BuyBy {
+  PSC = 'pcs',
+  KGS = 'kgs',
+  BOX = 'box',
+  PACK = 'pack',
+}
