+++ src/types/product/deliveryTypes.ts
@@ -1,0 +1,5 @@
+export interface IDelivery {
+  timeframe: number;
+  cost: number;
+  area: string[];
+}
