+++ src/types/product/descriptionTypes.ts
@@ -1,0 +1,5 @@
+export interface IDescription {
+  short: string;
+  long: string;
+  full: string;
+}
