+++ src/types/product/priceTypes.ts
@@ -1,0 +1,6 @@
+export interface IPrice {
+  amount: number;
+  discountedAmount: number;
+  discount: number;
+  currency: string;
+}
