+++ src/types/product/stockTypes.ts
@@ -1,0 +1,6 @@
+import { BuyBy } from './buyByTypes';
+
+export interface IStock {
+  amount: number;
+  measure: BuyBy;
+}
